Move some tests around to centralize them
