+++ src/common.rs
@@ -1786,7 +1786,7 @@ impl CpuRefreshKind {
 /// let mut system = System::new();
 ///
 /// // We don't want to update all memories information.
-/// system.refresh_memory_specifics(CpuRefreshKind::new().with_ram());
+/// system.refresh_memory_specifics(MemoryRefreshKind::new().with_ram());
 ///
 /// println!("total RAM: {}", system.total_memory());
 /// println!("free RAM:  {}", system.free_memory());
@@ -3773,66 +3773,5 @@ impl Cpu {
     /// ```
     pub fn frequency(&self) -> u64 {
         self.inner.frequency()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{MacAddr, ProcessStatus};
-
-    // This test only exists to ensure that the `Display` and `Debug` traits are implemented on the
-    // `ProcessStatus` enum on all targets.
-    #[test]
-    fn check_display_impl_process_status() {
-        println!("{} {:?}", ProcessStatus::Parked, ProcessStatus::Idle);
-    }
-
-    // Ensure that the `Display` and `Debug` traits are implemented on the `MacAddr` struct
-    #[test]
-    fn check_display_impl_mac_address() {
-        println!(
-            "{} {:?}",
-            MacAddr([0x1, 0x2, 0x3, 0x4, 0x5, 0x6]),
-            MacAddr([0xa, 0xb, 0xc, 0xd, 0xe, 0xf])
-        );
-    }
-
-    #[test]
-    fn check_mac_address_is_unspecified_true() {
-        assert!(MacAddr::UNSPECIFIED.is_unspecified());
-        assert!(MacAddr([0; 6]).is_unspecified());
-    }
-
-    #[test]
-    fn check_mac_address_is_unspecified_false() {
-        assert!(!MacAddr([1, 2, 3, 4, 5, 6]).is_unspecified());
-    }
-
-    // This test exists to ensure that the `TryFrom<usize>` and `FromStr` traits are implemented
-    // on `Uid`, `Gid` and `Pid`.
-    #[allow(clippy::unnecessary_fallible_conversions)]
-    #[test]
-    fn check_uid_gid_from_impls() {
-        use std::convert::TryFrom;
-        use std::str::FromStr;
-
-        #[cfg(not(windows))]
-        {
-            assert!(crate::Uid::try_from(0usize).is_ok());
-            assert!(crate::Uid::from_str("0").is_ok());
-        }
-        #[cfg(windows)]
-        {
-            assert!(crate::Uid::from_str("S-1-5-18").is_ok()); // SECURITY_LOCAL_SYSTEM_RID
-            assert!(crate::Uid::from_str("0").is_err());
-        }
-
-        assert!(crate::Gid::try_from(0usize).is_ok());
-        assert!(crate::Gid::from_str("0").is_ok());
-
-        assert!(crate::Pid::try_from(0usize).is_ok());
-        // If it doesn't panic, it's fine.
-        let _ = crate::Pid::from(0);
-        assert!(crate::Pid::from_str("0").is_ok());
     }
 }
